+++ src/main/java/CarView.java
@@ -10,40 +10,38 @@ import java.awt.event.ActionListener;
  * It initializes with being center on the screen and attaching it's controller in it's state.
  * It communicates with the Controller by calling methods of it when an action fires of in
  * each of it's components.
- * TODO: Write more actionListeners and wire the rest of the buttons
  **/
 
 
-// todo framtid: dela upp CarView i delar som {skapar fönstret}, {skapar knappar}, {talar med carC}, etc
 public class CarView extends JFrame{
-    private static final int X = 800;  //finns nu också i UserInput
-    private static final int Y = 800;  //finns nu också i UserInput
+    private static final int FrameSizeX = 800;  //finns nu också i UserInput
+    private static final int FrameSizeY = 800;  //finns nu också i UserInput
 
     // The controller member
     CarController carC;
 
-    DrawPanel drawPanel = new DrawPanel(X, Y-240);
+    DrawPanel drawPanel = new DrawPanel(FrameSizeX, FrameSizeY -240);
 
-    JPanel controlPanel = new JPanel();  //finns nu också i UserInput
+    ControlPanel controlPanel = new ControlPanel();  //finns nu också i UserInput
 
 
     //finns nu också i UserInput
-    JPanel gasPanel = new JPanel();
-    JSpinner gasSpinner = new JSpinner();
-    int gasAmount = 0;
-    JLabel gasLabel = new JLabel("Amount of gas");
-
-
-
-    JButton gasButton = new JButton("Gas");
-    JButton brakeButton = new JButton("Brake");
-    JButton turboOnButton = new JButton("Saab Turbo on");
-    JButton turboOffButton = new JButton("Saab Turbo off");
-    JButton liftBedButton = new JButton("Cars.Scania Lift Bed");
-    JButton lowerBedButton = new JButton("Lower Lift Bed");
-
-    JButton startButton = new JButton("Start all cars");
-    JButton stopButton = new JButton("Stop all cars");
+//    JPanel gasPanel = new JPanel();
+//    JSpinner gasSpinner = new JSpinner();
+//    int gasAmount = 0;
+//    JLabel gasLabel = new JLabel("Amount of gas");
+//
+//
+//
+//    JButton gasButton = new JButton("Gas");
+//    JButton brakeButton = new JButton("Brake");
+//    JButton turboOnButton = new JButton("Saab Turbo on");
+//    JButton turboOffButton = new JButton("Saab Turbo off");
+//    JButton liftBedButton = new JButton("Cars.Scania Lift Bed");
+//    JButton lowerBedButton = new JButton("Lower Lift Bed");
+//
+//    JButton startButton = new JButton("Start all cars");
+//    JButton stopButton = new JButton("Stop all cars");
 
     // Constructor
     public CarView(String framename, CarController cc){
@@ -56,107 +54,108 @@ public class CarView extends JFrame{
     private void initComponents(String title) {
 
         this.setTitle(title);
-        this.setPreferredSize(new Dimension(X,Y));
+        this.setPreferredSize(new Dimension(FrameSizeX, FrameSizeY));
         this.setLayout(new FlowLayout(FlowLayout.LEFT, 0, 0));
 
         this.add(drawPanel);
-
-
-
-        SpinnerModel spinnerModel =
-                new SpinnerNumberModel(0, //initial value
-                        0, //min
-                        1, //max
-                        1);//step
-        gasSpinner = new JSpinner(spinnerModel);
-        gasSpinner.addChangeListener(new ChangeListener() {
-            public void stateChanged(ChangeEvent e) {
-                gasAmount = (int) ((JSpinner)e.getSource()).getValue();
-            }
-        });
-
-        gasPanel.setLayout(new BorderLayout());
-        gasPanel.add(gasLabel, BorderLayout.PAGE_START);
-        gasPanel.add(gasSpinner, BorderLayout.PAGE_END);
-
-        this.add(gasPanel);
-
-        controlPanel.setLayout(new GridLayout(2,4));
-
-        controlPanel.add(gasButton, 0);
-        controlPanel.add(turboOnButton, 1);
-        controlPanel.add(liftBedButton, 2);
-        controlPanel.add(brakeButton, 3);
-        controlPanel.add(turboOffButton, 4);
-        controlPanel.add(lowerBedButton, 5);
-        controlPanel.setPreferredSize(new Dimension((X/2)+4, 200));
         this.add(controlPanel);
-        controlPanel.setBackground(Color.CYAN);
 
 
-        startButton.setBackground(Color.blue);
-        startButton.setForeground(Color.green);
-        startButton.setPreferredSize(new Dimension(X/5-15,200));
-        this.add(startButton);
 
+//        SpinnerModel spinnerModel =
+//                new SpinnerNumberModel(0, //initial value
+//                        0, //min
+//                        1, //max
+//                        1);//step
+//        gasSpinner = new JSpinner(spinnerModel);
+//        gasSpinner.addChangeListener(new ChangeListener() {
+//            public void stateChanged(ChangeEvent e) {
+//                gasAmount = (int) ((JSpinner)e.getSource()).getValue();
+//            }
+//        });
+//
+//        gasPanel.setLayout(new BorderLayout());
+//        gasPanel.add(gasLabel, BorderLayout.PAGE_START);
+//        gasPanel.add(gasSpinner, BorderLayout.PAGE_END);
+//
+//        this.add(gasPanel);
 
-        stopButton.setBackground(Color.red);
-        stopButton.setForeground(Color.black);
-        stopButton.setPreferredSize(new Dimension(X/5-15,200));
-        this.add(stopButton);
-
-        // This actionListener is for the gas button only
-        gasButton.addActionListener(new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                carC.gas(gasAmount);
-            }
-        });
-
-        brakeButton.addActionListener(new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                int brakeAmount = gasAmount;
-                carC.brake(brakeAmount);
-            }
-        });
-        turboOnButton.addActionListener(new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                carC.turboOn();
-            }
-        });
-        turboOffButton.addActionListener(new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                carC.turboOff();
-            }
-        });
-        liftBedButton.addActionListener(new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                carC.liftBed();
-            }
-        });
-        lowerBedButton.addActionListener(new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                carC.lowerBed();
-            }
-        });
-
-        startButton.addActionListener(new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                carC.startEngine();
-            }
-        });
-        stopButton.addActionListener(new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                carC.stopEngine();
-            }
-        });
+//        controlPanel.setLayout(new GridLayout(2,4));
+//
+//        controlPanel.add(gasButton, 0);
+//        controlPanel.add(turboOnButton, 1);
+//        controlPanel.add(liftBedButton, 2);
+//        controlPanel.add(brakeButton, 3);
+//        controlPanel.add(turboOffButton, 4);
+//        controlPanel.add(lowerBedButton, 5);
+//        controlPanel.setPreferredSize(new Dimension((FrameSizeX /2)+4, 200));
+//        this.add(controlPanel);
+//        controlPanel.setBackground(Color.CYAN);
+//
+//
+//        startButton.setBackground(Color.blue);
+//        startButton.setForeground(Color.green);
+//        startButton.setPreferredSize(new Dimension(FrameSizeX /5-15,200));
+//        this.add(startButton);
+//
+//
+//        stopButton.setBackground(Color.red);
+//        stopButton.setForeground(Color.black);
+//        stopButton.setPreferredSize(new Dimension(FrameSizeX /5-15,200));
+//        this.add(stopButton);
+//
+//        // This actionListener is for the gas button only
+//        gasButton.addActionListener(new ActionListener() {
+//            @Override
+//            public void actionPerformed(ActionEvent e) {
+//                carC.gas(gasAmount);
+//            }
+//        });
+//
+//        brakeButton.addActionListener(new ActionListener() {
+//            @Override
+//            public void actionPerformed(ActionEvent e) {
+//                int brakeAmount = gasAmount;
+//                carC.brake(brakeAmount);
+//            }
+//        });
+//        turboOnButton.addActionListener(new ActionListener() {
+//            @Override
+//            public void actionPerformed(ActionEvent e) {
+//                carC.turboOn();
+//            }
+//        });
+//        turboOffButton.addActionListener(new ActionListener() {
+//            @Override
+//            public void actionPerformed(ActionEvent e) {
+//                carC.turboOff();
+//            }
+//        });
+//        liftBedButton.addActionListener(new ActionListener() {
+//            @Override
+//            public void actionPerformed(ActionEvent e) {
+//                carC.liftBed();
+//            }
+//        });
+//        lowerBedButton.addActionListener(new ActionListener() {
+//            @Override
+//            public void actionPerformed(ActionEvent e) {
+//                carC.lowerBed();
+//            }
+//        });
+//
+//        startButton.addActionListener(new ActionListener() {
+//            @Override
+//            public void actionPerformed(ActionEvent e) {
+//                carC.startEngine();
+//            }
+//        });
+//        stopButton.addActionListener(new ActionListener() {
+//            @Override
+//            public void actionPerformed(ActionEvent e) {
+//                carC.stopEngine();
+//            }
+//        });
 
         // Make the frame pack all it's components by respecting the sizes if possible.
         this.pack();
